Add script to create a "dev" package, and allow building against it
Collects all the needed data from a source and build directory,
which then can be used instead of a source directory in combination
with an installed Qt Creator, to build plugins.

On Windows and OS X the plugin can still only built in the same
mode (release or debug) as the used Qt Creator install.

Change-Id: I21119cc0681f1a5f657c969f5d1e7a23d69aedfe
Reviewed-by: Marco Benelli <marco.benelli@qt.io>
Reviewed-by: Alessandro Portale <alessandro.portale@theqtcompany.com>