QmlDesigner: Fix lookup of navigator icons
The version check was simply wrong.
The version of the used item might be higher then the version in the
item library. We can still use the icon.

I removed the version check completely. There is no reason to have different
icons for different versions. This just increases the complexity and
introduces bugs without any reason/use cases.

Change-Id: Ib6d2040490739ee903ffe8eb308cd942b96444aa
Reviewed-by: Tim Jenssen <tim.jenssen@theqtcompany.com>