QtQuick*Wizards: Do never show an empty version dropdown
Always list all minimum Qt version options and pick the lowest by
default.

The wizard will be listed with no platform set and needs to do
something then. Before it did show an empty combobox and behaved
oddly from that point on when no kit with a Qt >= 5.3 was found.

Now it lists all possible versions and offers to set up kits for
you via the kits page when no matching kit is available.

Change-Id: I978268a2b8f1dfcfc8640460d905652abe91f1ae
Reviewed-by: Christian Stenger <christian.stenger@theqtcompany.com>