Add QtVersionManager::unsortedVersions()
To sort the list, we need to know the version number of all qt
versions. That requires running the qmake. Since in most cases
we don't need the sorted list at startup, add a method for that
and use it.

This prevents almost all calls to queryQmake for me, except for
a few for Android Qt Versions.

Change-Id: I6db89f214d68d07fee4b4dd15ee14c10d1248565
Reviewed-by: Orgad Shaneh <orgads@gmail.com>
Reviewed-by: Tobias Hunger <tobias.hunger@theqtcompany.com>