QmlDesigner: Move positioning using Positioners
Move positioning using Positioners to another context menu categeory.
The fact that we have Layouts and Positioners in Qt Quick and both do
something quite similar is something we cannot change.
I also do not like to remove Positioners from the designer, because
they have use cases. But having both in the same context menu was a mistake
I think.
So I moved the Positioner operations into their own menu/category.

The plan is to add a toolbar (eventually), but only support layouts in
the toolbar to minimize confusion.

Change-Id: I5a6abebca209eb43cac7404e12aa998c466cd62d
Reviewed-by: Tim Jenssen <tim.jenssen@theqtcompany.com>