QmakeProjectImporter: Slightly tweak spec handling
The removed code had actually no effect since it changed a variable that
was not used later. Since we are looking for a kit that has a identical
spec, there's no need to pass any spec in additional arguments.

Also on creating a temporary kit, don't set mkspec if it's the same
as the qt versions mkspec.

Change-Id: Ia37317a5b6fefa603106c8761279577271f32a7e
Reviewed-by: Tobias Hunger <tobias.hunger@theqtcompany.com>