iOS: Fix kit creation with Xcode 7
Xcode 7's platform plist does not specify a concrete version
for the SDK in the SDK name anymore.
So, if we do not find an SDK with the same name, look for an SDK
with a version that matches the platform's 'Version' setting.

Trying to figure out from the settings looks inherently fragile
to me though, and we might be better off just using
xcrun -sdk <sdk> --show-sdk-path

Change-Id: Ief45f03aa7cd5193f6c24b087eb635d3f5ba5298
Reviewed-by: Daniel Teske <daniel.teske@theqtcompany.com>