Do not show detailed progress information by default.
The transition phase from having it in the mode bar is long over, so
people that were used to always seeing the detailed view it will either
have adapted, or have it set in their settings.
Not showing the details is the saner default setting, since the popup
does and will always hide other UI like the editor scrollbar, buttons in
find tool bar, content of output panes, etc etc.

Change-Id: Ib34dee976c46b5d19b81cebcca416612b673fb29
Task-number: QTCREATORBUG-9837
Reviewed-by: Orgad Shaneh <orgads@gmail.com>