Utils: Make the meaning of PathChooser::changed() consistent.
This signal was emitted from two places. In one of them, the argument
was the pure string from the line edit, and in the other one the "raw
path", that is, the aforementioned string processed by cleanPath() and
fromNativeSeparators().
We now use the latter in both contexts. A look at the slots does not
suggest anyone is interested in the other value rather than this one.

Change-Id: Iacb6e0815004b2ef0107bfef03b97b46015da989
Reviewed-by: Eike Ziller <eike.ziller@theqtcompany.com>