Qbs: Adapt rpath handling to how it is normally implemented on OS X.
When using rpaths on OS X, the install name prefix should always be set
to "@rpath" and the rpaths should be appropriately set to the locations
in which that binary looks for its dependencies.

While this change is behaviorally identical, it brings the project
setup in line with how developers expect these properties to be set, and
thus reduces confusion.

As an aside, it now correctly follows the principle of relocatability
(install name prefixes like @rpath/Frameworks are only partially
relocatable as opposed to @rpath alone which is fully relocatable).

Change-Id: I7b5cc7a316d2608037ea9ab087029baa85aee2ba
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>