Timeline: Optimize selection tools for fewer scene graph rebuilds
The selection details "windows" contain frequently updated text and can
be moved around in the scene. We don't want their contents to be batched
with the rest of the scene as there may be many events in the timeline
and rebuilding the batch they are in is expensive. The clip property has
the side effect of preventing nodes within the clipped area from getting
batched with nodes outside it. Thus, by setting the clip property on
the details we reduce the number of batch rebuilds for the actual
timeline.

Having a transparent rectangle in there doesn't help the renderer,
either, so we replace it with a plain item.

Change-Id: Ie6bac1324d3bc6b8b3d56e658a766f1f90be9046
Task-number: QTCREATORBUG-14983
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>