FakeVim: Fix caps lock in macros.
This fixes the bug that while recording a macrom, hitting caps lock
(scroll lock, num lock) inserts some key into the buffer (for caps '$')
so when you replay the macro you get incorrect output.

Other keys that are not in vimKeyNames may behave similarly but It is
important that macros support those three (especially caps lock).

Change-Id: I2854266f86a22e9d38fb1c6c5d923b11a56a5995
Reviewed-by: hjk <hjk@theqtcompany.com>