Fix infinite QWaitCondition:wait() in discardFile*()
The original idea was that we would passively wait for another thread to
clean up the locker, hence the check-sleep-loop. This was all dandy,
except for *also* using the wait condition: this was a) mostly pointless
(it would just avoid a few iterations of the wait loop) and b) buggy (if
there were no other waiting threads, the actual reader thread wouldn't
know that it needs to wake somebody up).
As the passive waiting is ugly, we instead fix the use of the wait
condition, and do away with the loop.

Task-number: QTCREATORBUG-15181
Change-Id: I477dbe7cda49ceca9aa387910d94ad763a43012b
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>
Reviewed-by: Daniel Teske <daniel.teske@theqtcompany.com>