Show folders as tree view for resources
This patch builds a tree view of the paths to files
in resource files.
Previously all files were shown in a flat structure
under prefixes with the entire path as part of their
displayed name.
Because folders based on prefixes are different from
folders in the paths, a new class, SimpleResourceFolderNode,
was introduced.

Change-Id: Ifc4773cff6a678b50e64b0d56713f80704e12f6f
Reviewed-by: Svenn-Arne Dragly <s@dragly.com>
Reviewed-by: Tobias Hunger <tobias.hunger@theqtcompany.com>