Locator: Do not loose focus on active window change
We cleared focus on the locator input when it lost focus because the
active window changed, so the popup is not reopened when the Qt Creator
window is activated again. Instead keep the focus in the locator input,
but do not open the popup if the focus comes back to it because of an
active window change.
Then we need to make sure that the popup is always opened when
explicitly requested through show(...).

Change-Id: Icd78673bece542ae78340b2ab6aa7834d700a3b7
Task-number: QTCREATORBUG-13548
Task-number: QTCREATORBUG-9953
Reviewed-by: André Hartmann <aha_1980@gmx.de>