QmlDesigner: Fix crash
We removed support for non default properties in the navigator.
But the model still allows QML code like this:  toolBar: ToolBar {...}

In this case ToolBar is not visible in the navigator. But the ToolBar
is available/visible in the form editor. Reparenting anything to ToolBar did
crash. Eventually we have to re enable non default properties, but this
patch only fixes the crash.

Task-number: QTCREATORBUG-15111
Change-Id: Ifa037975c99762fb8067afffa118332044842d7f
Reviewed-by: Eike Ziller <eike.ziller@theqtcompany.com>
Reviewed-by: Tim Jenssen <tim.jenssen@theqtcompany.com>