QmlDebug: Avoid dangerous interactions with QPacketProtocol
After deleteLater() a QObject does have another chance to send more
signals. So, if we do that with the underlying QIODevice, we should
disconnect it from the debug connection (which then doesn't have a
QPacketProtocol anymore). Also, setting dev to 0 in QPacketProtocol is
an incredibly bad idea, as we never check it for 0. The only reason
this piece of code is never triggered is that we never generate
invalid packets on purpose.

Change-Id: I80e7f494bf55fa1a510b6f0bdd50cf45cae23734
Task-number: QTCREATORBUG-15496
Reviewed-by: Joerg Bornemann <joerg.bornemann@theqtcompany.com>