report errors from projects included by cumulative evaluation
so far, we would suppress any errors which occur during the (exact)
evaluation of projects which were included only by cumulative evaluation
of a parent project (see f53d37d48). this makes sense, as excluded
projects are likely to fail, producing unsettling noise on the way.
however, having no indication at all why a subproject failed to load is
not helpful, either. so print these errors, but tag them with a prefix.

note that we continue to suppress messages from the cumulative
evaluation of all projects, including from the exact evaluation of .prf
files done even in this mode. this is done because this output is
expected to be quite noisy and unhelpful. however, it can be argued that
this isn't a wise choice for projects which are known to likely fail for
reasons which would not impact our ability to show (an approximation of)
the project in the project explorer. for the time being, we assume that
the expected errors occur later on (typically during the .prf evaluation
phase due to unknown QT entries), so an unexpected error which would
also affect the cumulative evaluation is displayed nonetheless.

Change-Id: Idc2675744169b7cb52c0542a80075850e2b9317c
Reviewed-by: Tobias Hunger <tobias.hunger@theqtcompany.com>