ProjectExplorer: Introduce Runnable and Connection concepts.
A 'Runnable' contains (possibly target-specific) data to describe
something that could be run (e.g. a remote executable with
arguments, working directory and environment). A 'Connection'
contains (possibly target-specific) data to describe a connection
to target that will be used to run a Runnable.

Runnable and Connection objects are used in the target
RunControlFactories and are opaque to the project explorer.

Change-Id: I2293dffa71148646cfb52c9b582ccd2675b145a8
Reviewed-by: Tobias Hunger <tobias.hunger@theqtcompany.com>