ProjectExplorer: Let a RunConfiguration specify a Runnable.
The (target specific) RunConfiguationFactory can specify what it considers
a Runnable for the target. The (also target specific) RunControlFactory
can use that either unmodified or tweak it to set up a RunControl,
(and finally, target specific code runs it).

So this still keeps the target code in full control on what is passed
where, but reduces the need for boilerplate in the cases where
the produced and consumed Runnables are the same (or similar).

Change-Id: Ied59430462f4a45ad0006a99c1efd8b07cd181b2
Reviewed-by: Tobias Hunger <tobias.hunger@theqtcompany.com>