QbsProjectManager: Remove "dry run" build option.
The way this option works is not IDE-compatible, since the build process
changes the build graph and relies on the changes being transient simply
by not storing the build graph afterwards. This is fine for the command-
line tool, but Qt Creator keeps the build graph open and subsequent
"real" builds will finish immediately even when nothing has actually
been built yet.
Perhaps it would be possible to do the implementation differently, but
having the "dry run" option in Qt Creator is not valuable enough to
justify that effort.

Change-Id: Ic99ddef63555f6029c5857d2cfd8dc48d8a72914
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@theqtcompany.com>
Reviewed-by: Jake Petroules <jake.petroules@theqtcompany.com>